Configure Github-native Dependabot for stylebot and source-reminder
This commit adds the configuration for these two directories.  Didn't
add hoedown-helper because we don't have a requirements.txt for that app.

Left a comment for future additions to this repo, discussing how to Configure
future directories.  One challenge is how to configure dependabot.yml
to look in multiple directories for npm dependencies, for example.

Believe a workaround would be to add another code block for each package manager/directory,
until Github allows us to give a list value for directories.