Build SourceReminder dependencies to local directory
This commit edits the build step to make sure
dependencies are build locally to the directory and
are present when being packaged and uploaded to Lambda.

Right now this creates a messy directory, so I .gitignored
the installed dependencies and created files, and started
to add the steps for moving to AWS SAM CLI for deploys
which'll be cleaner.