Adjust the range of dates we're comparing
We filter the list of entries from the sheet to narrow down our
comparison to pub dates within the start and end date. In the sheet,
the pub date would be something like 10/1/2021, but in the code
our start date has a leading zero with 10/01/2021. Pub dates without
the leading zero like the one in this case weren't being referenced
and the reminder was reporting those stories as missing.
To account for this we just widened the range to an extra day
before the start and after the end.