Deploy "latest" image on "master" branch changes
This commit updates the action from pushing a new AWS ECR image on each commit, using the commit sha, to only build and deploy a "latest" image to ECR on changes to the master branch.

Source reminder for Lambda will read from the "latest" image, so this keeps ECR Registry cleaner by just having single image in the registry.  Since this will be autodeploying based on the latest change to master branch, we can just use git for versioning (always can revert a commit/pr and this will build and deploy that image as "latest").