Update docker image layer order to improve cache
Docker caches layers based on the order, so the least likely layers to be changed should be higher in the docker layer stack.  I think the code for the lambda function is more likely to be changed than the python dependencies, so this commit moves the function code lower in the stack (so we don't reinstall python packages each time we just make code changes).