Don't check incoming data event for slack token
This commit removes logic to check the incoming data body for a "token"
key, which is being checked against the known SLACK_BOT_TOKEN environment variable.

The purpose of this block is for when slack bot lambda functions are
triggered by an event in slack (like hoedown helper), to confirm that the
incoming event is actually coming from Slack like we expect.

This doesn't apply to source-reminder, since the function is not triggered
by slack events, but just on a schedule -- it should be able to run with an
empty event data argument.  This commit fixes that.