Configure Google Service Credentials as base64 environment variable
This commit changes how we receive the Google Service credentials.
Instead of relying on a locally co-located json file (which is hard to
do when we configure this function to run as a lambda container), we
expect the json file to be a base64-encoded string.

This allows the json to be stored as an environment variable (locally and
in lambda).  To use this environment variable in the lambda function,
we decode the base64 string, then serialize the string to json to be
used in the authorization logic.

Commit also adds the needed environment variable to docker-env-example.