Add Github Action step to sync Lambda function with newest Image
I tried locally pushing new images to AWS ECR, and found that
the lambda function wasn't automatically picking up the newest :latest
tagged image, I had to click "Deploy New Image."  This commit adds an automated
step to the Continuous deployment that will update the lambda function to
always use the newest / most recent github action-pushed image in the
next function invokation.

Note:  from the SourceReminder Lambda configuration page in AWS, I added
the necessary IAM role as a resource-based permission so that the IAM role
used in this Github repo secret/Github Action has the permission to run the
"aws lambda update-function-code" api call.

I also switched the image tag that we are pushing to AWS ECR from :latest
to use the github commit hash.  I noticed that there were a lot of dangling/untagged
images after pushing a few with the same :latest tag.  Figured that this
is a little easier to follow since you can always reference the SHA to look at
exactly which commit is in the docker image being invoked.

Found the method of doing this in this tutorial:
https://shpals.medium.com/create-aws-lambda-from-ecr-docker-image-and-integrate-it-with-github-ci-cd-pipeline-dfa3015b5ee0